requirements.txt
numpy==1.22.1
pandas==1.4.3
scikit-image==0.19.3
scikit-learn==1.0.2
tensorflow==2.9.1
joblib==1.1.0
pillow==9.2.0
streamlit==1.15.1
opencv_python==4.6.0.66